Added handling of lack of connection
New exception class added. In the constructor, an amount of attempts to try and connect if there is an error connecting to the Bittrex host can be specified, in addition to how long to wait between each attempt. If the amount of attempts is exceeded, the (unchecked) exception is thrown, and can be handled as required.